Add support for switching to sub-camera
Add functionality to switch between main and sub-cameras, such as switching from the main (rear) camera to the sub (front) camera.

To use this feature, set the sub-camera name in the Sub Device Name field of SimpleCamera before starting the camera. This can be configured via the inspector or changed at runtime.